+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 18.0.0
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Add techdivision/import-cli-simple#216
+* Switch to latest techdivision/import-product-ee 18.* and techdivision/import-product-bundle 18.* version as dependency
+
 # Version 17.0.0
 
 ## Bugfixes